requirements.txt
asgiref==3.10.0
boto3==1.40.72
botocore==1.40.72
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
jmespath==1.0.1
PyJWT==2.10.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
